# openapi-specs/openapi.json
{
    "x-generator": "NSwag v14.6.2.0 (NJsonSchema v11.5.2.0 (Newtonsoft.Json v13.0.0.0))",
    "openapi": "3.0.0",
    "info": {
        "title": "Dating API",
        "description": "An ASP.NET Core Web API for managing Dating items",
        "termsOfService": "https://example.com/terms",
        "contact": {
            "name": "Example Contact",
            "url": "https://example.com/contact"
        },
        "license": {
            "name": "Example License",
            "url": "https://example.com/license"
        },
        "version": "v1"
    },
    "paths": {
        "/api/Account/register": {
            "post": {
                "tags": [
                    "Account"
                ],
                "summary": "Creates an User.",
                "description": "Sample request:\n            \n    POST api/account/register\n    {\n       \"displayName\": \"displayName\",\n       \"email\": \"test@test.com\",\n       \"password\": \"password\"\n    }",
                "operationId": "Account_Register",
                "requestBody": {
                    "x-name": "request",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterRequest"
                            }
                        }
                    },
                    "required": true,
                    "x-position": 1
                },
                "responses": {
                    "200": {
                        "description": "Returns the newly created item",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "If the item is null",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/Account/login": {
            "post": {
                "tags": [
                    "Account"
                ],
                "summary": "Login as an User.",
                "operationId": "Account_Login",
                "requestBody": {
                    "x-name": "request",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    },
                    "required": true,
                    "x-position": 1
                },
                "responses": {
                    "200": {
                        "description": "Returns the valid user token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "If the username or password are incorrect",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/Error/bad-request": {
            "get": {
                "tags": [
                    "Error"
                ],
                "operationId": "Error_GetBadRequest",
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/octet-stream": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/Error/auth": {
            "get": {
                "tags": [
                    "Error"
                ],
                "operationId": "Error_GetAuth",
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/octet-stream": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/Error/not-found": {
            "get": {
                "tags": [
                    "Error"
                ],
                "operationId": "Error_GetNotFound",
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/octet-stream": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/Error/server-error": {
            "get": {
                "tags": [
                    "Error"
                ],
                "operationId": "Error_GetServerError",
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/octet-stream": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/Members": {
            "get": {
                "tags": [
                    "Members"
                ],
                "operationId": "Members_GetMembers",
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Member"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Members"
                ],
                "operationId": "Members_UpdateMember",
                "requestBody": {
                    "x-name": "request",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MemberUpdateRequest"
                            }
                        }
                    },
                    "required": true,
                    "x-position": 1
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/octet-stream": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/Members/{id}": {
            "get": {
                "tags": [
                    "Members"
                ],
                "operationId": "Members_GetMember",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "x-position": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Member"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/Members/{id}/photos": {
            "get": {
                "tags": [
                    "Members"
                ],
                "operationId": "Members_GetPhotos",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "x-position": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Photo"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/Members/photo": {
            "post": {
                "tags": [
                    "Members"
                ],
                "operationId": "Members_AddPhoto",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "ContentType": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "ContentDisposition": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "Headers": {
                                        "type": "array",
                                        "nullable": true,
                                        "items": {}
                                    },
                                    "Length": {
                                        "type": "integer",
                                        "format": "int64"
                                    },
                                    "Name": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "FileName": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Photo"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/Members/photo/{photoId}": {
            "put": {
                "tags": [
                    "Members"
                ],
                "operationId": "Members_SetMainPhoto",
                "parameters": [
                    {
                        "name": "photoId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        },
                        "x-position": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/octet-stream": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "UserResponse": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "displayName": {
                        "type": "string"
                    },
                    "imageUrl": {
                        "type": "string",
                        "nullable": true
                    },
                    "token": {
                        "type": "string"
                    }
                }
            },
            "ProblemDetails": {
                "type": "object",
                "additionalProperties": {
                    "nullable": true
                },
                "properties": {
                    "type": {
                        "type": "string",
                        "nullable": true
                    },
                    "title": {
                        "type": "string",
                        "nullable": true
                    },
                    "status": {
                        "type": "integer",
                        "format": "int32",
                        "nullable": true
                    },
                    "detail": {
                        "type": "string",
                        "nullable": true
                    },
                    "instance": {
                        "type": "string",
                        "nullable": true
                    }
                }
            },
            "RegisterRequest": {
                "type": "object",
                "additionalProperties": false,
                "required": [
                    "displayName",
                    "email",
                    "password"
                ],
                "properties": {
                    "displayName": {
                        "type": "string",
                        "minLength": 1
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "minLength": 1
                    },
                    "password": {
                        "type": "string",
                        "minLength": 6
                    }
                }
            },
            "LoginRequest": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                }
            },
            "Member": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "birthDay": {
                        "type": "string",
                        "format": "date"
                    },
                    "imageUrl": {
                        "type": "string",
                        "nullable": true
                    },
                    "displayName": {
                        "type": "string"
                    },
                    "created": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "lastActive": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "gender": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "city": {
                        "type": "string"
                    },
                    "country": {
                        "type": "string"
                    }
                }
            },
            "Photo": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "url": {
                        "type": "string"
                    },
                    "publicId": {
                        "type": "string",
                        "nullable": true
                    },
                    "memberId": {
                        "type": "string"
                    }
                }
            },
            "MemberUpdateRequest": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                    "displayName": {
                        "type": "string",
                        "nullable": true
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "city": {
                        "type": "string",
                        "nullable": true
                    },
                    "country": {
                        "type": "string",
                        "nullable": true
                    }
                }
            }
        }
    }
}